+++ src/pyramid/util.py
@@ -667,3 +667,11 @@ def is_unbound_method(fn):
             return True
 
     return False
+
+
+def reraise(tp, value, tb=None):
+    if value is None:
+        value = tp
+    if value.__traceback__ is not tb:
+        raise value.with_traceback(tb)
+    raise value
